Support cross-running 32-bit efibootmgr build on 64-bit kernel
To comply UEFI spec, build from the existing implementation
can run on a kernel with the same architecture only, either
64-bit or 32-bit.

This patch provides support to run a 32-bit build on a 64-bit
kernel when the new compiling option is enabled and the kernel
does not have support for efivarfs.

When this option is not enabled, the build is compiled in the
same way of the existing. It is not effective either in 64-bit
build.

Note: We cannot test efibootmgr with the option enabled in adb
shell on a 64-bit kernel in Android build so far. There is no EFI
service exposed from BIOS in legacy mode.

Change-Id: Ie9ed0638e1a43c0826c1a607e46374e27b45ded0
Signed-off-by: Radivoje Jovanovic <radivoje.jovanovic@intel.com>
Reviewed-on: https://android.intel.com/172750
Reviewed-by: Dattatraya Kulkarni, Anisha <anisha.dattatraya.kulkarni@intel.com>
Tested-by: Dattatraya Kulkarni, Anisha <anisha.dattatraya.kulkarni@intel.com>
Reviewed-by: Uyyala, Sridhar <sridhar.uyyala@intel.com>